requirements.txt
pandas
python-dotenv
langchain-openai
streamlit
pybind11>=2.12
langchain-experimental
tabulate
llama-index-llms-gemini
llama-index 
llama-index-experimental